fix(search): change elliptical search item avatars to circular avatars (#669)
This is kind of a workaround to add the falsely worded commit.
"style(search): change elliptical search item avatars to circular
avatars" to the generated changelog. I mistook style commits for UI
style, but they are meant for code style commits. I didn't want to
bother rewording / rebasing the commit history, so this is the fix to
include the message...